allow connection pool timeouts to be sub-second - change `ConnPoolConfig.TimeoutSeconds` to `float64` - parse `-queryserver-config-query-pool-timeout` and `-queryserver-config-txpool-timeout as float64`
vitess maintains three connection pools http://et.sy/9vlndg7
- `OltpReadPool` is used for `SELECT`s when `OLTP` mode is in use
- `OlapReadPool` is used for `SELECT`s when `OLAP` mode is in use
- `TxPool` is used for `INSERT`s/`UPDATE`s and transactions when `OLTP` mode is in use

the `-queryserver-config-query-pool-timeout` flag maps to `OltpReadPool.TimeoutSeconds` and controls
how long vttablet waits for a connection from the `OltpReadPool`

the `-queryserver-config-txpool-timeout` flag maps to `TxPool.TimeoutSeconds` and controls
how long vttablet waits for a connection from the `TxPool`

in `v6.0.20` there is no equivalent flag for for the `OlapReadPool`, I also tested in `4.0.1` and the
`-queryserver-config-query-pool-timeout` flag has no effect for OLAP mode. A look at the source code
also confirms it never applied http://et.sy/1soa6nq

+++ go/vt/vttablet/tabletserver/tabletenv/config.go
@@ -97,8 +97,8 @@ func init() {
 	flag.IntVar(&currentConfig.QueryCacheSize, "queryserver-config-query-cache-size", defaultConfig.QueryCacheSize, "query server query cache size, maximum number of queries to be cached. vttablet analyzes every incoming query and generate a query plan, these plans are being cached in a lru cache. This config controls the capacity of the lru cache.")
 	flag.IntVar(&currentConfig.SchemaReloadIntervalSeconds, "queryserver-config-schema-reload-time", defaultConfig.SchemaReloadIntervalSeconds, "query server schema reload time, how often vttablet reloads schemas from underlying MySQL instance in seconds. vttablet keeps table schemas in its own memory and periodically refreshes it from MySQL. This config controls the reload time.")
 	flag.IntVar(&currentConfig.Oltp.QueryTimeoutSeconds, "queryserver-config-query-timeout", defaultConfig.Oltp.QueryTimeoutSeconds, "query server query timeout (in seconds), this is the query timeout in vttablet side. If a query takes more than this timeout, it will be killed.")
-	flag.IntVar(&currentConfig.OltpReadPool.TimeoutSeconds, "queryserver-config-query-pool-timeout", defaultConfig.OltpReadPool.TimeoutSeconds, "query server query pool timeout (in seconds), it is how long vttablet waits for a connection from the query pool. If set to 0 (default) then the overall query timeout is used instead.")
-	flag.IntVar(&currentConfig.TxPool.TimeoutSeconds, "queryserver-config-txpool-timeout", defaultConfig.TxPool.TimeoutSeconds, "query server transaction pool timeout, it is how long vttablet waits if tx pool is full")
+	flag.Float64Var(&currentConfig.OltpReadPool.TimeoutSeconds, "queryserver-config-query-pool-timeout", defaultConfig.OltpReadPool.TimeoutSeconds, "query server query pool timeout (in seconds), it is how long vttablet waits for a connection from the query pool. If set to 0 (default) then the overall query timeout is used instead.")
+	flag.Float64Var(&currentConfig.TxPool.TimeoutSeconds, "queryserver-config-txpool-timeout", defaultConfig.TxPool.TimeoutSeconds, "query server transaction pool timeout, it is how long vttablet waits if tx pool is full")
 	flag.IntVar(&currentConfig.OltpReadPool.IdleTimeoutSeconds, "queryserver-config-idle-timeout", defaultConfig.OltpReadPool.IdleTimeoutSeconds, "query server idle timeout (in seconds), vttablet manages various mysql connection pools. This config means if a connection has not been used in given idle timeout, this connection will be removed from pool. This effectively manages number of connection objects and optimize the pool performance.")
 	flag.IntVar(&currentConfig.OltpReadPool.MaxWaiters, "queryserver-config-query-pool-waiter-cap", defaultConfig.OltpReadPool.MaxWaiters, "query server query pool waiter limit, this is the maximum number of queries that can be queued waiting to get a connection")
 	flag.IntVar(&currentConfig.TxPool.MaxWaiters, "queryserver-config-txpool-waiter-cap", defaultConfig.TxPool.MaxWaiters, "query server transaction pool waiter limit, this is the maximum number of transactions that can be queued waiting to get a connection")
@@ -227,11 +227,11 @@ type TabletConfig struct {
 
 // ConnPoolConfig contains the config for a conn pool.
 type ConnPoolConfig struct {
-	Size               int `json:"size,omitempty"`
-	TimeoutSeconds     int `json:"timeoutSeconds,omitempty"`
-	IdleTimeoutSeconds int `json:"idleTimeoutSeconds,omitempty"`
-	PrefillParallelism int `json:"prefillParallelism,omitempty"`
-	MaxWaiters         int `json:"maxWaiters,omitempty"`
+	Size               int     `json:"size,omitempty"`
+	TimeoutSeconds     float64 `json:"timeoutSeconds,omitempty"`
+	IdleTimeoutSeconds int     `json:"idleTimeoutSeconds,omitempty"`
+	PrefillParallelism int     `json:"prefillParallelism,omitempty"`
+	MaxWaiters         int     `json:"maxWaiters,omitempty"`
 }
 
 // OltpConfig contains the config for oltp settings.
